+++ src/explore-education-statistics-admin/src/contexts/EditingContext.tsx
@@ -15,6 +15,8 @@ export interface UnSavedEdit {
 export interface EditingContextState {
   isEditing: boolean;
   setEditing: (isEditing: boolean) => void;
+  isTablePreview: boolean;
+  setTablePreview: (isTablePreview: boolean) => void;
   unSavedEdits: UnSavedEdit[];
   setUnSavedEdits: (unSavedEdits: UnSavedEdit[]) => void;
 }
@@ -22,6 +24,8 @@ export interface EditingContextState {
 export const EditingContext = createContext<EditingContextState>({
   isEditing: false,
   setEditing: noop,
+  isTablePreview: false,
+  setTablePreview: noop,
   unSavedEdits: [],
   setUnSavedEdits: noop,
 });
@@ -30,10 +34,11 @@ export function useEditingContext() {
   return useContext(EditingContext);
 }
 
-interface EditingContextProviderProps {
+export interface EditingContextProviderProps {
   children: ReactNode | ((state: EditingContextState) => ReactNode);
   value?: {
     isEditing: boolean;
+    isTablePreview?: boolean;
     unSavedEdits?: UnSavedEdit[];
   };
 }
@@ -42,19 +47,23 @@ export const EditingContextProvider = ({
   children,
   value = {
     isEditing: true,
+    isTablePreview: false,
   },
 }: EditingContextProviderProps) => {
   const [isEditing, setEditing] = useState<boolean>(value.isEditing);
+  const [isTablePreview, setTablePreview] = useState<boolean>(false);
   const [unSavedEdits, setUnSavedEdits] = useState<UnSavedEdit[]>([]);
 
   const state = useMemo<EditingContextState>(() => {
     return {
       isEditing,
       setEditing,
+      isTablePreview,
+      setTablePreview,
       unSavedEdits,
       setUnSavedEdits,
     };
-  }, [isEditing, unSavedEdits]);
+  }, [isEditing, isTablePreview, unSavedEdits]);
 
   return (
     <EditingContext.Provider value={state}>
