+++ src/explore-education-statistics-admin/src/pages/release/content/components/ReleasePreviewTableToolFinalStep.tsx
@@ -1,29 +1,22 @@
-import Link from '@admin/components/Link';
-import { preReleaseContentRoute } from '@admin/routes/preReleaseRoutes';
-import { ReleaseRouteParams } from '@admin/routes/releaseRoutes';
 import TableHeadersForm from '@common/modules/table-tool/components/TableHeadersForm';
 import TimePeriodDataTable from '@common/modules/table-tool/components/TimePeriodDataTable';
 import { FullTable } from '@common/modules/table-tool/types/fullTable';
 import { TableHeadersConfig } from '@common/modules/table-tool/types/tableHeaders';
+import { Publication } from '@common/services/publicationService';
 import DownloadTable from '@common/modules/table-tool/components/DownloadTable';
 import TableToolInfo from '@common/modules/table-tool/components/TableToolInfo';
-import { BasicPublicationDetails } from '@admin/services/publicationService';
-import React, { memo, useEffect, useRef, useState } from 'react';
-import { generatePath } from 'react-router-dom';
+import React, { useEffect, useRef, useState } from 'react';
 
-interface TableToolFinalStepProps {
-  publication?: BasicPublicationDetails;
-  releaseId: string;
+interface ReleasePreviewTableToolFinalStepProps {
+  publication?: Publication;
   table: FullTable;
   tableHeaders: TableHeadersConfig;
 }
-
-const PreReleaseTableToolFinalStep = ({
+const ReleasePreviewTableToolFinalStep = ({
   publication,
-  releaseId,
   table,
   tableHeaders,
-}: TableToolFinalStepProps) => {
+}: ReleasePreviewTableToolFinalStepProps) => {
   const dataTableRef = useRef<HTMLElement>(null);
   const [currentTableHeaders, setCurrentTableHeaders] = useState<
     TableHeadersConfig
@@ -32,6 +25,20 @@ const PreReleaseTableToolFinalStep = ({
   useEffect(() => {
     setCurrentTableHeaders(tableHeaders);
   }, [tableHeaders]);
+
+  const getMethodologyLink = () => {
+    if (publication?.methodology) {
+      return publication.methodology.title;
+    }
+    if (publication?.externalMethodology) {
+      return (
+        <a href={publication.externalMethodology.url}>
+          {publication.externalMethodology.title}
+        </a>
+      );
+    }
+    return null;
+  };
 
   return (
     <div className="govuk-!-margin-bottom-4">
@@ -55,7 +62,6 @@ const PreReleaseTableToolFinalStep = ({
           tableHeadersConfig={currentTableHeaders}
         />
       )}
-
       {publication && table && (
         <>
           <DownloadTable
@@ -66,25 +72,9 @@ const PreReleaseTableToolFinalStep = ({
 
           <TableToolInfo
             table={table}
-            contactDetails={{
-              contactName: publication?.contact?.contactName || '',
-              contactTelNo: publication?.contact?.contactTelNo || '',
-              teamEmail: publication?.contact?.teamEmail || '',
-              teamName: publication?.contact?.teamName || '',
-            }}
-            releaseLink={
-              <Link
-                to={generatePath<ReleaseRouteParams>(
-                  preReleaseContentRoute.path,
-                  {
-                    publicationId: publication.id,
-                    releaseId,
-                  },
-                )}
-              >
-                {publication.title}
-              </Link>
-            }
+            contactDetails={publication.contact}
+            methodologyLink={getMethodologyLink()}
+            releaseLink={<span>{publication.title}</span>}
           />
         </>
       )}
@@ -92,4 +82,4 @@ const PreReleaseTableToolFinalStep = ({
   );
 };
 
-export default memo(PreReleaseTableToolFinalStep);
+export default ReleasePreviewTableToolFinalStep;
