+++ src/explore-education-statistics-admin/src/pages/release/content/contexts/__tests__/ReleaseContentContext.test.tsx
@@ -3,14 +3,13 @@ import {
   releaseReducer as originalReleaseReducer,
 } from '@admin/pages/release/content/contexts/ReleaseContentContext';
 import { ReleaseDispatchAction } from '@admin/pages/release/content/contexts/ReleaseContentContextActionTypes';
-import { EditableRelease } from '@admin/services/releaseContentService';
 import {
   Comment,
   EditableBlock,
   EditableContentBlock,
 } from '@admin/services/types/content';
-import { ReleaseType } from '@common/services/publicationService';
 import { DataBlock, Table } from '@common/services/types/blocks';
+import { testEditableRelease } from '@admin/pages/release/__data__/testEditableRelease';
 import { produce } from 'immer';
 
 const emptyTable: Table = {
@@ -21,291 +20,6 @@ const emptyTable: Table = {
     rowGroups: [],
     rows: [],
   },
-};
-
-const basicRelease: EditableRelease = {
-  id: '6a97c9b6-eaa2-4d22-7ba9-08d7bec1ba1a',
-  title: 'Academic Year 2020/21',
-  yearTitle: '2020/21',
-  coverageTitle: 'Academic Year',
-  releaseName: '2020',
-  slug: '2020-21',
-  publicationId: '8586c490-7027-4a8b-9df5-cf105c3e88ec',
-  approvalStatus: 'Draft',
-  dataLastPublished: '',
-  published: '',
-  publication: {
-    id: '8586c490-7027-4a8b-9df5-cf105c3e88ec',
-    title: 'My Pub',
-    slug: 'my-pub',
-    otherReleases: [],
-    legacyReleases: [],
-    dataSource: '',
-    description: '',
-    summary: '',
-    topic: { theme: { title: 'Children, early years and social care' } },
-    contact: {
-      teamName: 'Explore Education Statistics',
-      teamEmail: 'explore.statistics@education.gov.uk',
-      contactName: 'Cameron Race',
-      contactTelNo: '07780991976',
-    },
-    methodology: {
-      id: '8807af2b-d9a7-4efa-a376-08d7b451a98e',
-      title: 'First methodology',
-      slug: '',
-      summary: '',
-    },
-  },
-  latestRelease: false,
-  type: {
-    id: '9d333457-9132-4e55-ae78-c55cb3673d7c',
-    title: ReleaseType.OfficialStatistics,
-  },
-  updates: [
-    {
-      id: '262cf6c8-db96-40d8-8fb1-b55028a9f55b',
-      reason: 'First published',
-      releaseId: '',
-      on: new Date(),
-    },
-  ],
-  content: [
-    {
-      id: '30e57106-66e2-43a3-fc07-08d7bf65f293',
-      order: 0,
-      heading: 'New section 3',
-      content: [
-        {
-          id: 'eb809a91-af3d-438d-632c-08d7c408aca5',
-          body: '',
-          type: 'MarkDownBlock',
-          order: 0,
-          comments: [
-            {
-              id: 'comment-1',
-              content: 'A comment',
-              createdBy: {
-                id: 'user-1',
-                firstName: 'Bau1',
-                lastName: '',
-                email: 'bau1@test.com',
-              },
-              created: '2020-03-09T09:39:53.736',
-            },
-            {
-              id: 'comment-2',
-              content: 'another comment',
-              createdBy: {
-                id: 'user-1',
-                firstName: 'Bau1',
-                lastName: '',
-                email: 'bau1@test.com',
-              },
-              created: '2020-03-09T09:40:16.534',
-            },
-          ],
-        },
-        {
-          heading: "Table showing 'prma' from 'My Pub' in England for 2017/18",
-          name: 'DataBlock 1',
-          source: '',
-          query: {
-            subjectId: '36aa28ce-83ca-49b5-8c27-b34e77b062c9',
-            timePeriod: {
-              startYear: 2017,
-              startCode: 'AY',
-              endYear: 2017,
-              endCode: 'AY',
-            },
-            filters: ['5a1b6c93-c9c3-4710-f71f-08d7bec1ddb3'],
-            indicators: [
-              '272f4b12-9dfe-45ef-fdf9-08d7bec1dd71',
-              'd314e2a9-5069-4d79-fdfa-08d7bec1dd71',
-              '214f5fb8-c12b-411e-fdfb-08d7bec1dd71',
-              '99f5bbc3-de9a-4831-fdf8-08d7bec1dd71',
-            ],
-            locations: {
-              country: ['E92000001'],
-            },
-            includeGeoJson: true,
-          },
-          charts: [],
-          table: emptyTable,
-          type: 'DataBlock',
-          id: '69a9522d-501d-441a-9ee5-260ede5cd85c',
-          order: 1,
-          comments: [],
-        },
-      ],
-    },
-    {
-      id: '47b7cabf-4855-44b8-5fdc-08d7c1e45526',
-      order: 1,
-      heading: 'New section',
-      content: [],
-    },
-  ],
-  summarySection: {
-    id: '6662107c-5c58-4ace-d1b1-08d7bec1ba1b',
-    order: 0,
-    content: [],
-    heading: '',
-  },
-  headlinesSection: {
-    id: '25a6cfbc-2cb5-43da-d1b7-08d7bec1ba1b',
-    heading: '',
-    order: 0,
-    content: [
-      {
-        id: 'f27bf214-a244-41a7-b09b-08d7c1e10b96',
-        type: 'MarkDownBlock',
-        body: '',
-        order: 0,
-        comments: [],
-      },
-    ],
-  },
-  keyStatisticsSection: {
-    id: '10e39b93-0304-43fb-d1b3-08d7bec1ba1b',
-    order: 0,
-    heading: '',
-    content: [
-      {
-        heading:
-          "Table showing Main reason for issue: Unauthorised family holiday absence for 'prma' from 'My Pub' in England for 2017/18",
-        name: 'Holiday (Key Stat)',
-        source: '',
-        query: {
-          subjectId: '36aa28ce-83ca-49b5-8c27-b34e77b062c9',
-          timePeriod: {
-            startYear: 2017,
-            startCode: 'AY',
-            endYear: 2017,
-            endCode: 'AY',
-          },
-          filters: ['5a1b6c93-c9c3-4710-f71f-08d7bec1ddb3'],
-          indicators: ['272f4b12-9dfe-45ef-fdf9-08d7bec1dd71'],
-          locations: {
-            country: ['E92000001'],
-          },
-          includeGeoJson: true,
-        },
-        charts: [],
-        table: emptyTable,
-        type: 'DataBlock',
-        id: 'e2db1389-8220-41f0-a29a-bb8dd1ccfc9c',
-        order: 0,
-        comments: [],
-      },
-      {
-        heading:
-          "Table showing Main reason for issue: Arriving late for 'prma' from 'My Pub' in England for 2017/18",
-        name: 'Main Reason (Key Stat)',
-        source: '',
-        query: {
-          subjectId: '36aa28ce-83ca-49b5-8c27-b34e77b062c9',
-          timePeriod: {
-            startYear: 2017,
-            startCode: 'AY',
-            endYear: 2017,
-            endCode: 'AY',
-          },
-          filters: ['5a1b6c93-c9c3-4710-f71f-08d7bec1ddb3'],
-          indicators: ['d314e2a9-5069-4d79-fdfa-08d7bec1dd71'],
-          locations: {
-            country: ['E92000001'],
-          },
-          includeGeoJson: true,
-        },
-        charts: [],
-        table: emptyTable,
-        type: 'DataBlock',
-        id: 'aa9f1b63-abb3-41a2-bd57-fcf24dfd71ed',
-        order: 1,
-        comments: [],
-      },
-      {
-        heading:
-          "Table showing Main reason for issue: Absence due to other unauthorised circumstances for 'prma' from 'My Pub' in England for 2017/18",
-        name: "Unauth'd (Key Stat)",
-        source: '',
-        query: {
-          subjectId: '36aa28ce-83ca-49b5-8c27-b34e77b062c9',
-          timePeriod: {
-            startYear: 2017,
-            startCode: 'AY',
-            endYear: 2017,
-            endCode: 'AY',
-          },
-          filters: ['5a1b6c93-c9c3-4710-f71f-08d7bec1ddb3'],
-          indicators: ['214f5fb8-c12b-411e-fdfb-08d7bec1dd71'],
-          locations: {
-            country: ['E92000001'],
-          },
-          includeGeoJson: true,
-        },
-        charts: [],
-        table: emptyTable,
-        type: 'DataBlock',
-        id: 'ba1ff405-1f44-4509-9dcf-5822df5b6f8c',
-        order: 2,
-        comments: [],
-      },
-    ],
-  },
-  keyStatisticsSecondarySection: {
-    heading: '',
-    id: '5509024d-9389-40d1-d1b5-08d7bec1ba1b',
-    order: 0,
-    content: [
-      {
-        heading: "Table showing 'prma' from 'My Pub' in England for 2017/18",
-        name: 'Aggregate table (Secondary)',
-        source: '',
-        query: {
-          subjectId: '36aa28ce-83ca-49b5-8c27-b34e77b062c9',
-          timePeriod: {
-            startYear: 2017,
-            startCode: 'AY',
-            endYear: 2017,
-            endCode: 'AY',
-          },
-          filters: ['5a1b6c93-c9c3-4710-f71f-08d7bec1ddb3'],
-          indicators: [
-            '272f4b12-9dfe-45ef-fdf9-08d7bec1dd71',
-            'd314e2a9-5069-4d79-fdfa-08d7bec1dd71',
-            '214f5fb8-c12b-411e-fdfb-08d7bec1dd71',
-          ],
-          locations: {
-            country: ['E92000001'],
-          },
-          includeGeoJson: true,
-        },
-        charts: [],
-        table: emptyTable,
-        type: 'DataBlock',
-        id: 'a3197018-66b6-4ce5-97fa-da2355270c40',
-        order: 0,
-        comments: [],
-      },
-    ],
-  },
-  downloadFiles: [
-    {
-      id: 'download-1',
-      extension: 'csv',
-      fileName: 'prma.csv',
-      name: 'prma',
-      size: '268 Kb',
-      type: 'Data',
-    },
-  ],
-  hasPreReleaseAccessList: false,
-  hasMetaGuidance: true,
-  publishScheduled: '2020-03-03T00:00:00',
-  nextReleaseDate: { year: 2020, month: 3, day: 4 },
-  relatedInformation: [],
 };
 
 const basicDataBlock: DataBlock = {
@@ -335,7 +49,7 @@ describe('ReleaseContentContext', () => {
     expect(
       releaseReducer(
         {
-          release: basicRelease,
+          release: testEditableRelease,
           canUpdateRelease: false,
           availableDataBlocks: [],
           unresolvedComments: [],
@@ -346,7 +60,7 @@ describe('ReleaseContentContext', () => {
         },
       ),
     ).toEqual({
-      release: basicRelease,
+      release: testEditableRelease,
       canUpdateRelease: false,
       availableDataBlocks: [basicDataBlock],
       unresolvedComments: [],
@@ -355,13 +69,13 @@ describe('ReleaseContentContext', () => {
 
   test('REMOVE_BLOCK_FROM_SECTION removes a block from a section', () => {
     const sectionKey = 'keyStatisticsSection';
-    const keyStatsSection = basicRelease[sectionKey];
+    const keyStatsSection = testEditableRelease[sectionKey];
     const removingBlockId = keyStatsSection.content[0].id;
     const originalLength = keyStatsSection.content.length;
 
     const { release } = releaseReducer(
       {
-        release: basicRelease,
+        release: testEditableRelease,
         canUpdateRelease: true,
         availableDataBlocks: [basicDataBlock],
         unresolvedComments: [],
@@ -391,7 +105,7 @@ describe('ReleaseContentContext', () => {
 
   test('REMOVE_BLOCK_FROM_SECTION removes a block from content section', () => {
     const sectionKey = 'content';
-    const releaseContent = basicRelease[sectionKey];
+    const releaseContent = testEditableRelease[sectionKey];
     const removingBlockId = (releaseContent[0]
       .content as EditableContentBlock[])[0].id;
     const originalLength = (releaseContent[0].content as EditableContentBlock[])
@@ -399,7 +113,7 @@ describe('ReleaseContentContext', () => {
 
     const { release } = releaseReducer(
       {
-        release: basicRelease,
+        release: testEditableRelease,
         canUpdateRelease: true,
         availableDataBlocks: [basicDataBlock],
         unresolvedComments: [],
@@ -427,14 +141,14 @@ describe('ReleaseContentContext', () => {
 
   test('UPDATE_BLOCK_FROM_SECTION updates a block from section', () => {
     const sectionKey = 'headlinesSection';
-    const section = basicRelease[sectionKey];
+    const section = testEditableRelease[sectionKey];
     const blockToUpdate = (section.content as EditableContentBlock[])[0];
 
     const newBody = 'This is some updated text!';
 
     const { release } = releaseReducer(
       {
-        release: basicRelease,
+        release: testEditableRelease,
         canUpdateRelease: true,
         availableDataBlocks: [basicDataBlock],
         unresolvedComments: [],
@@ -462,15 +176,15 @@ describe('ReleaseContentContext', () => {
 
   test('UPDATE_BLOCK_FROM_SECTION updates a block from a content section', () => {
     const sectionKey = 'content';
-    const sectionId = basicRelease.content[0].id;
-    const blockToUpdate = (basicRelease.content[0]
+    const sectionId = testEditableRelease.content[0].id;
+    const blockToUpdate = (testEditableRelease.content[0]
       .content as EditableContentBlock[])[0];
 
     const newBody = 'This is some updated text!';
 
     const { release } = releaseReducer(
       {
-        release: basicRelease,
+        release: testEditableRelease,
         canUpdateRelease: true,
         availableDataBlocks: [basicDataBlock],
         unresolvedComments: [],
@@ -498,7 +212,7 @@ describe('ReleaseContentContext', () => {
 
   test('ADD_BLOCK_TO_SECTION adds a block to a section', () => {
     const sectionKey = 'summarySection';
-    const section = basicRelease[sectionKey];
+    const section = testEditableRelease[sectionKey];
     const newBlock: EditableContentBlock = {
       id: '123',
       order: 0,
@@ -511,7 +225,7 @@ describe('ReleaseContentContext', () => {
 
     const { release } = releaseReducer(
       {
-        release: basicRelease,
+        release: testEditableRelease,
         canUpdateRelease: false,
         availableDataBlocks: [],
         unresolvedComments: [],
@@ -539,7 +253,7 @@ describe('ReleaseContentContext', () => {
 
   test('ADD_BLOCK_TO_SECTION adds a block to a content section', () => {
     const sectionKey = 'content';
-    const section = basicRelease.content[0];
+    const section = testEditableRelease.content[0];
 
     const newBlock: EditableContentBlock = {
       id: '123',
@@ -553,7 +267,7 @@ describe('ReleaseContentContext', () => {
 
     const { release } = releaseReducer(
       {
-        release: basicRelease,
+        release: testEditableRelease,
         canUpdateRelease: true,
         availableDataBlocks: [basicDataBlock],
         unresolvedComments: [],
@@ -647,7 +361,7 @@ describe('ReleaseContentContext', () => {
 
     const { release } = releaseReducer(
       {
-        release: basicRelease,
+        release: testEditableRelease,
         canUpdateRelease: true,
         availableDataBlocks: [basicDataBlock],
         unresolvedComments: [],
@@ -656,7 +370,7 @@ describe('ReleaseContentContext', () => {
         type: 'UPDATE_SECTION_CONTENT',
         payload: {
           meta: {
-            sectionId: basicRelease.content[0].id,
+            sectionId: testEditableRelease.content[0].id,
             sectionKey: 'content',
           },
           sectionContent: newContent,
@@ -669,10 +383,10 @@ describe('ReleaseContentContext', () => {
   });
 
   test('ADD_CONTENT_SECTION adds a new section to release content', () => {
-    const originalLength = basicRelease.content.length;
+    const originalLength = testEditableRelease.content.length;
     const { release } = releaseReducer(
       {
-        release: basicRelease,
+        release: testEditableRelease,
         canUpdateRelease: true,
         availableDataBlocks: [basicDataBlock],
         unresolvedComments: [],
@@ -683,7 +397,7 @@ describe('ReleaseContentContext', () => {
           section: {
             heading: 'A new section',
             id: 'new-section-1',
-            order: basicRelease.content.length,
+            order: testEditableRelease.content.length,
             content: [],
           },
         },
@@ -696,8 +410,8 @@ describe('ReleaseContentContext', () => {
   });
 
   test("SET_CONTENT sets the release's content", () => {
-    const contentSectionsLength = basicRelease.content.length;
-    const contentSectionsReversed = basicRelease.content.map(
+    const contentSectionsLength = testEditableRelease.content.length;
+    const contentSectionsReversed = testEditableRelease.content.map(
       (section, index) => {
         return { ...section, order: contentSectionsLength - (1 + index) };
       },
@@ -705,7 +419,7 @@ describe('ReleaseContentContext', () => {
 
     const { release } = releaseReducer(
       {
-        release: basicRelease,
+        release: testEditableRelease,
         canUpdateRelease: true,
         availableDataBlocks: [basicDataBlock],
         unresolvedComments: [],
@@ -724,10 +438,10 @@ describe('ReleaseContentContext', () => {
   });
 
   test('UPDATE_CONTENT_SECTION updates a content section', () => {
-    const basicSection = basicRelease.content[0];
+    const basicSection = testEditableRelease.content[0];
     const { release } = releaseReducer(
       {
-        release: basicRelease,
+        release: testEditableRelease,
         canUpdateRelease: true,
         availableDataBlocks: [basicDataBlock],
         unresolvedComments: [],
@@ -764,13 +478,13 @@ describe('ReleaseContentContext', () => {
     };
 
     const updatedComments = [
-      ...basicRelease.content[0].content[0].comments,
+      ...testEditableRelease.content[0].content[0].comments,
       newComment,
     ];
 
     const { release, unresolvedComments } = releaseReducer(
       {
-        release: basicRelease,
+        release: testEditableRelease,
         canUpdateRelease: true,
         availableDataBlocks: [basicDataBlock],
         unresolvedComments: [],
@@ -779,8 +493,8 @@ describe('ReleaseContentContext', () => {
         type: 'UPDATE_BLOCK_COMMENTS',
         payload: {
           meta: {
-            sectionId: basicRelease.content[0].id,
-            blockId: basicRelease.content[0].content[0].id,
+            sectionId: testEditableRelease.content[0].id,
+            blockId: testEditableRelease.content[0].content[0].id,
             sectionKey,
           },
           comments: [...updatedComments],
