+++ src/explore-education-statistics-admin/src/pages/release/datablocks/ReleaseTableToolPage.tsx
@@ -3,90 +3,22 @@ import {
   releaseDataBlocksRoute,
   ReleaseRouteParams,
 } from '@admin/routes/releaseRoutes';
+import releaseContentService from '@admin/services/releaseContentService';
 import LoadingSpinner from '@common/components/LoadingSpinner';
+import ReleasePreviewTableTool from '@admin/pages/release/content/components/ReleasePreviewTableTool';
 import useAsyncHandledRetry from '@common/hooks/useAsyncHandledRetry';
-import TableHeadersForm from '@common/modules/table-tool/components/TableHeadersForm';
-import TableToolWizard, {
-  InitialTableToolState,
-} from '@common/modules/table-tool/components/TableToolWizard';
-import TimePeriodDataTable from '@common/modules/table-tool/components/TimePeriodDataTable';
-import WizardStep from '@common/modules/table-tool/components/WizardStep';
-import WizardStepHeading from '@common/modules/table-tool/components/WizardStepHeading';
-import { FullTable } from '@common/modules/table-tool/types/fullTable';
-import { TableHeadersConfig } from '@common/modules/table-tool/types/tableHeaders';
-import tableBuilderService from '@common/services/tableBuilderService';
-import React, { useEffect, useRef, useState } from 'react';
+import React from 'react';
 import { generatePath, RouteComponentProps } from 'react-router-dom';
-
-interface ReleaseTableToolFinalStepProps {
-  table: FullTable;
-  tableHeaders: TableHeadersConfig;
-}
-
-const ReleaseTableToolFinalStep = ({
-  table,
-  tableHeaders,
-}: ReleaseTableToolFinalStepProps) => {
-  const dataTableRef = useRef<HTMLElement>(null);
-  const [currentTableHeaders, setCurrentTableHeaders] = useState<
-    TableHeadersConfig
-  >();
-
-  useEffect(() => {
-    setCurrentTableHeaders(tableHeaders);
-  }, [tableHeaders]);
-
-  return (
-    <div className="govuk-!-margin-bottom-4">
-      <TableHeadersForm
-        initialValues={currentTableHeaders}
-        onSubmit={tableHeaderConfig => {
-          setCurrentTableHeaders(tableHeaderConfig);
-
-          if (dataTableRef.current) {
-            dataTableRef.current.scrollIntoView({
-              behavior: 'smooth',
-              block: 'start',
-            });
-          }
-        }}
-      />
-      {table && currentTableHeaders && (
-        <TimePeriodDataTable
-          ref={dataTableRef}
-          fullTable={table}
-          tableHeadersConfig={currentTableHeaders}
-        />
-      )}
-    </div>
-  );
-};
 
 const ReleaseTableToolPage = ({
   match,
 }: RouteComponentProps<ReleaseRouteParams>) => {
-  const { publicationId, releaseId } = match.params;
+  const { releaseId, publicationId } = match.params;
 
-  const { value: initialState, isLoading } = useAsyncHandledRetry<
-    InitialTableToolState | undefined
-  >(async () => {
-    const {
-      subjects,
-    } = await tableBuilderService.getReleaseSubjectsAndHighlights(releaseId);
-
-    return {
-      initialStep: 1,
-      subjects,
-      query: {
-        publicationId,
-        releaseId,
-        subjectId: '',
-        indicators: [],
-        filters: [],
-        locations: {},
-      },
-    };
-  }, [releaseId]);
+  const { value, isLoading } = useAsyncHandledRetry(
+    () => releaseContentService.getContent(releaseId),
+    [releaseId],
+  );
 
   return (
     <>
@@ -100,36 +32,12 @@ const ReleaseTableToolPage = ({
       >
         Back
       </Link>
-
       <LoadingSpinner loading={isLoading}>
-        {initialState && (
-          <>
-            <h2>Table tool</h2>
-
-            <TableToolWizard
-              themeMeta={[]}
-              hidePublicationSelectionStage
-              initialState={initialState}
-              finalStep={({ response, query }) => (
-                <WizardStep>
-                  {wizardStepProps => (
-                    <>
-                      <WizardStepHeading {...wizardStepProps}>
-                        Explore data
-                      </WizardStepHeading>
-
-                      {query && response && (
-                        <ReleaseTableToolFinalStep
-                          table={response.table}
-                          tableHeaders={response.tableHeaders}
-                        />
-                      )}
-                    </>
-                  )}
-                </WizardStep>
-              )}
-            />
-          </>
+        {value && (
+          <ReleasePreviewTableTool
+            releaseId={releaseId}
+            publication={value.release.publication}
+          />
         )}
       </LoadingSpinner>
     </>
