+++ src/explore-education-statistics-common/src/modules/table-tool/components/TableToolInfo.tsx
@@ -1,0 +1,53 @@
+import { FullTable } from '@common/modules/table-tool/types/fullTable';
+import Accordion from '@common/components/Accordion';
+import AccordionSection from '@common/components/AccordionSection';
+import React, { ReactNode } from 'react';
+import { PublicationContact } from '@common/services/publicationService';
+
+interface Props {
+  table?: FullTable;
+  contactDetails?: PublicationContact;
+  methodologyLink?: ReactNode;
+  releaseLink?: ReactNode;
+}
+
+const TableToolInfo = ({
+  table,
+  contactDetails,
+  methodologyLink,
+  releaseLink,
+}: Props) => {
+  return (
+    <Accordion id="TableToolInfo">
+      {table && (
+        <AccordionSection heading="Related information">
+          <ul className="govuk-list">
+            {releaseLink && <li>Publication: {releaseLink}</li>}
+            {methodologyLink && <li>Methodology: {methodologyLink}</li>}
+          </ul>
+        </AccordionSection>
+      )}
+      {contactDetails && (
+        <AccordionSection heading="Contact us">
+          <p>
+            If you have a question about the data or methods used to create this
+            table contact the named statistician:
+          </p>
+          <h4 className="govuk-heading-s govuk-!-font-weight-bold">
+            {contactDetails.teamName}
+          </h4>
+          <p>Named statistician: {contactDetails.contactName}</p>
+          <p>
+            Email:{' '}
+            <a href={`mailto:${contactDetails.teamEmail}`}>
+              {contactDetails.teamEmail}
+            </a>
+          </p>
+          <p>Telephone: {contactDetails.contactTelNo}</p>
+        </AccordionSection>
+      )}
+    </Accordion>
+  );
+};
+
+export default TableToolInfo;
